+++ src/lib/types/ucs2-string.ts
@@ -10,7 +10,6 @@ import { Lazy, VersionedType } from "../types";
 
 export type Name = "ucs2-string";
 export const name: Name = "ucs2-string";
-export type T = string;
 export namespace json {
   export type Input = string;
   export type Output = string;
@@ -106,7 +105,7 @@ export interface Options {
  * PS: This type does not deal with Unicdoe normalization either. Use CodepointString and CodepointArray if you need
  * it.
  */
-export class Ucs2StringType implements VersionedType<T, json.Input, json.Output, Diff> {
+export class Ucs2StringType implements VersionedType<string, json.Input, json.Output, Diff> {
   readonly name: Name = name;
   readonly allowUnicodeRegExp: boolean;
   readonly pattern?: RegExp;
@@ -166,11 +165,11 @@ export class Ucs2StringType implements VersionedType<T, json.Input, json.Output,
     return jsonType;
   }
 
-  readTrustedJson(input: json.Output): T {
+  readTrustedJson(input: json.Output): string {
     return input;
   }
 
-  readJson(input: any): T {
+  readJson(input: any): string {
     const error: Error | undefined = this.testError(input);
     if (error !== undefined) {
       throw error;
@@ -178,11 +177,11 @@ export class Ucs2StringType implements VersionedType<T, json.Input, json.Output,
     return input;
   }
 
-  writeJson(val: T): json.Output {
+  writeJson(val: string): json.Output {
     return val;
   }
 
-  testError(val: T): Error | undefined {
+  testError(val: string): Error | undefined {
     if (typeof val !== "string") {
       return WrongTypeError.create("string", val);
     }
@@ -217,23 +216,23 @@ export class Ucs2StringType implements VersionedType<T, json.Input, json.Output,
     return undefined;
   }
 
-  test(val: T): boolean {
+  test(val: string): boolean {
     return this.testError(val) === undefined;
   }
 
-  equals(val1: T, val2: T): boolean {
+  equals(val1: string, val2: string): boolean {
     return val1 === val2;
   }
 
-  clone(val: T): T {
+  clone(val: string): string {
     return val;
   }
 
-  diff(oldVal: T, newVal: T): Diff | undefined {
+  diff(oldVal: string, newVal: string): Diff | undefined {
     return oldVal === newVal ? undefined : [oldVal, newVal];
   }
 
-  patch(oldVal: T, diff: Diff | undefined): T {
+  patch(oldVal: string, diff: Diff | undefined): string {
     return diff === undefined ? oldVal : diff[1];
   }
 
@@ -267,5 +266,3 @@ export class Ucs2StringType implements VersionedType<T, json.Input, json.Output,
     Object.freeze(this);
   }
 }
-
-export { Ucs2StringType as Type };